+++ packages/logger/tsconfig.build.json
@@ -1,4 +1,4 @@
 {
   "extends": "./tsconfig.json",
-  "exclude": ["./src/**/*spec.ts"]
+  "exclude": ["test"]
 }
