+++ packages/nestjs-module/package.json
@@ -25,7 +25,9 @@
     "lib": "lib"
   },
   "scripts": {
-    "build": "tsc -b tsconfig.build.json",
+    "prebuild": "rimraf lib",
+    "build": "tsc -p tsconfig.build.json",
+    "postbuild": "mv ./lib/src/* ./lib && rmdir lib/src",
     "test": "jest",
     "test:cov": "jest --coverage"
   },
