+++ packages/platform-fastify/package.json
@@ -26,7 +26,9 @@
     "url": "git+https://github.com/jmcdo29/ogma.git"
   },
   "scripts": {
-    "build": "tsc -b tsconfig.build.json",
+    "prebuild": "rimraf lib",
+    "build": "tsc -p tsconfig.build.json",
+    "postbuild": "mv ./lib/src/* ./lib && rmdir lib/src",
     "test": "jest",
     "test:cov": "jest --coverage"
   },
