+++ packages/platform-graphql-fastify/tsconfig.json
@@ -1,13 +1,8 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "rootDir": "./src",
+    "rootDir": "./",
     "outDir": "./lib"
   },
-  "include": ["./src"],
-  "references": [
-    {
-      "path": "../logger"
-    }
-  ]
+  "include": ["./src", "./test"]
 }
