+++ packages/platform-grpc/tsconfig.build.json
@@ -1,4 +1,4 @@
 {
   "extends": "./tsconfig.json",
-  "exclude": ["test", "./src/**/*spec.ts"]
+  "exclude": ["test"]
 }
