+++ packages/platform-tcp/tsconfig.json
@@ -1,13 +1,8 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "rootDir": "./src",
+    "rootDir": "./",
     "outDir": "./lib"
   },
-  "include": ["./src"],
-  "references": [
-    {
-      "path": "../logger"
-    }
-  ]
+  "include": ["./src", "./test"]
 }
