+++ tsconfig.json
@@ -1,6 +1,5 @@
 {
   "compilerOptions": {
-    "composite": true,
     "module": "commonjs",
     "declaration": true,
     "removeComments": false,
